refactor: CircularProgressDialogFragment dialog 패키지로 이동
