+++ app/src/main/java/com/stopsmoke/kekkek/presentation/dialog/CircularProgressDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.stopsmoke.kekkek.presentation.progress
+package com.stopsmoke.kekkek.presentation.dialog
 
 import android.app.Dialog
 import android.os.Bundle
