+++ app/src/main/res/layout/fragment_circular_progress_dialog.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.progress.CircularProgressDialogFragment">
+    tools:context=".presentation.dialog.CircularProgressDialogFragment">
 
     <com.google.android.material.progressindicator.CircularProgressIndicator
         app:indicatorColor="@color/primary_blue"
